[atlona] Make volume a whole number & remove StringUtils (#9631)

+++ bundles/org.openhab.binding.atlona/src/main/java/org/openhab/binding/atlona/internal/pro3/AtlonaPro3Handler.java
@@ -323,7 +323,7 @@ public class AtlonaPro3Handler extends AtlonaHandler<AtlonaPro3Capabilities> {
                             break;
                         case AtlonaPro3Constants.CHANNEL_VOLUME:
                             if (command instanceof DecimalType) {
-                                final double level = ((DecimalType) command).doubleValue();
+                                final int level = ((DecimalType) command).intValue();
                                 atlonaHandler.setVolume(portNbr, level);
                             } else {
                                 logger.debug("Received a VOLUME channel command with a non DecimalType: {}", command);
